p0_backend/pyproject.toml
[build-system]
requires = ["poetry-core>=1.6.1"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "p0-backend"
version = "1.0.0"
description = "Protocol0 ableton remote script backend"
readme = "README.md"
authors = ["Thibault Lebrun <thibaultlebrun@live.fr>"]
license = "MIT"
repository = "https://github.com/lebrunthibault/code"
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Development Status :: 4 - Beta"
]
keywords = ["remote script", "ableton"]

[tool.poetry.dependencies]
python = "3.11.3"
apispec=" ^5.2.2"
asyncclick=" ^8.1.3.2"
click=" ^8.1.3"
keyboard=" ^0.13.5"
Pillow=" ^9.5.0"
psutil=" ^5.9.1"
PyAutoGUI=" ^0.9.53"
pydantic=" ^1.9.1"
PySimpleGUI=" ^4.60.5"
pywin32=" ^304"
six=" ^1.16.0"
uvicorn = {extras = ["standard"], version = "^0.22.0"}
p0-script = {path = "C:/Users/thiba/dev/code/projects/p0_script", develop = true}
loguru = "^0.7.0"
celery = "^5.2.7"
redis = "^4.5.5"
mido = "^1.2.10"
pygame = "^2.4.0"
fastapi = "^0.96.0"
python-dotenv = "^1.0.0"
rx = "^3.2.0"
requests = "^2.31.0"
tzdata = "^2023.3"
openapi-spec-validator = "^0.5.6"

[tool.poetry.dev-dependencies]
flake8 = "*"
vulture = "*"
pytest = "*"
watchdog = {extras = ["watchmedo"], version = "^3.0.0"}

[tool.mypy]
python_version = 3.10
exclude = ['venv', 'tests', 'vulture_whitelist.py']
pretty = false
ignore_missing_imports = true
follow_imports = 'skip'
# cannot remove misc because of _Framework. union-attr possibly
disable_error_code = ['misc','union-attr', 'override', 'var-annotated']
show_error_codes = true
warn_unreachable = true
strict_equality = true
disallow_any_explicit = true
disallow_untyped_calls = false
disallow_untyped_defs = false
disallow_incomplete_defs = false
disallow_untyped_decorators = false
incremental = false

[tool.black]
line-length = 100

[tool.vulture]
min_confidence = 60
make_whitelist = true
sort_by_size = true
verbose = false
exclude = [
    "venv/",
    "protocol0_stub",
    "sdk_generation/",
    "routes.py"
]
